+++ apps/web/src/features/shorts/api/shorts.api.ts
@@ -1,5 +1,5 @@
 import type { Shorts } from '@prisma/client';
-import { http } from '../shared/api/client';
+import { http } from '@/shared/api/client';
 
 export const fetchShorts = async ({ cursor, limit }: { cursor: string; limit: number }) => {
     const res = await http.get<Shorts[]>('/api/shorts', {
