+++ apps/web/src/features/shorts/server/index.ts
@@ -1,0 +1,3 @@
+export { ShortRepository } from './short.repository';
+export { ShortService } from './short.service';
+export { searchShorts } from './youtube.api';
